+++ integrations/armorcode/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.2.54 (2026-08-16)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.48.8
+
+
+## 0.2.53 (2026-08-13)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.48.7
+
+
 ## 0.2.52 (2026-08-13)
 
 
@@ -424,6 +440,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 0.2.0 (2026-06-15)
+
 
 ### Improvements
 
@@ -1470,6 +1487,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped starlette version to 0.49.3
 
+
 ## 0.1.10-beta (2025-11-09)
 
 
@@ -1552,12 +1570,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.0-dev (2025-10-06)
 
+
 ### Features
 
 - Created Armorcode Ocean integration with support for Product, Sub Product and Findings
 
 
 ## 0.1.0-beta (2025-10-09)
+
 
 ### Features