+++ integrations/aws/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.3.130 (2026-08-16)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.48.8
+
+
+## 0.3.129 (2026-08-13)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.48.7
+
+
 ## 0.3.128 (2026-08-13)
 
 
@@ -1283,7 +1299,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.2.228 (2025-12-25)
 
 
-## Bug Fixes
+### Bug Fixes
 
 - Fixed spec.yaml conventions
 
@@ -1322,10 +1338,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.2.223 (2025-12-08)
 
+
 ### Bug Fixes
 
 - Move account discovery to resync start. AWS account list is now refreshed before each resync cycle, fixing an issue where changes in AWS Organizations (e.g., added/removed accounts) weren't detected unless the integration was restarted. (Fixes PORT-16916)
-
 
 
 ## 0.2.222 (2025-12-16)
@@ -1552,6 +1568,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped starlette version to 0.49.3
 - Bumped aiohttp version to 3.13.2
 
+
 ## 0.2.194 (2025-11-09)
 
 
@@ -1590,6 +1607,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - Renaming from on-premise to self hosted
+
 
 ## 0.2.189 (2025-11-02)
 
@@ -1633,6 +1651,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.2.184-post (2025-10-21)
 
+
 ### Improvements
 
 - Deprecate aws v2 integration
@@ -1647,6 +1666,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 0.2.183 (2025-10-16)
+
 
 ### Improvements
 
@@ -2929,13 +2949,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped ocean version to ^0.9.12 (#1)
 
 
-# Port_Ocean 0.2.25 (2024-08-05)
+## 0.2.25 (2024-08-05)
+
 
 ### Improvements
 
 - Add live events error handling
 
-# Port_Ocean 0.2.24 (2024-08-05)
+
+## 0.2.24 (2024-08-05)
+
 
 ### Improvements
 
@@ -3023,6 +3046,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add access denied exception support (#1)
 
+
 ## 0.2.13 (2024-07-10)
 
 
@@ -3038,6 +3062,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix default useGetResourceAPI property name (#1)
 - Use by default the actual S3 Bucket region instead of default region used to fetch it (#2)
+
 
 ## 0.2.11 (2024-07-09)
 
@@ -3058,7 +3083,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.2.9 (2024-07-02)
 
 
-### Bugfix
+### Bug Fixes
 
 - Ensure default region for global resources (#1)
 
@@ -3094,12 +3119,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed default mapping to include describeResources for resources which doesn't include tags by default from cloud control (#1)
 
+
 ## 0.2.4 (2024-06-17)
 
 
 ### Improvements
 
 - Fix _aws_credentials overflow bug (#1)
+
 
 ## 0.2.3 (2024-06-17)
 
@@ -3117,6 +3144,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - Run all single describe in parallel (#1)
+
 
 ## 0.2.1 (2024-06-16)
 
@@ -3140,7 +3168,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - Bumped ocean version to ^0.8.0 (#1)
-
 
 
 ## 0.1.7 (2024-06-13)